banyak

+++ backend/controllers/superadmin/AuthItemController.php
@@ -59,6 +59,9 @@ class AuthItemController extends Controller
      */
     public function actionView($id)
     {
+         if(!Yii::$app->user->can('view-akses'))
+             throw new ForbiddenHttpException;
+
         return $this->render('view', [
             'model' => $this->findModel($id),
         ]);
@@ -71,6 +74,9 @@ class AuthItemController extends Controller
      */
     public function actionCreate()
     {
+        if(!Yii::$app->user->can('create-akses'))
+             throw new ForbiddenHttpException;
+
         $model = new AuthItemModel();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -95,6 +101,9 @@ class AuthItemController extends Controller
      */
     public function actionUpdate($id)
     {
+        if(!Yii::$app->user->can('update-akses'))
+             throw new ForbiddenHttpException;
+
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -119,6 +128,9 @@ class AuthItemController extends Controller
      */
     public function actionDelete($id)
     {
+        if(!Yii::$app->user->can('delete-akses'))
+             throw new ForbiddenHttpException;
+
         if($this->findModel($id)->delete()){
             Yii::$app->getSession()->setFlash('success', [
                     'message' => "Data Berhasil Dihapus",
